ai_service/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sentence-transformers==3.4.1
numpy==2.2.6
pypdf==5.9.0
python-docx==1.2.0
